Project Hydrate V3 (Final)/requirements.txt
python==3.11.3
pillow==9.5.0
pip==23.1.2
future==0.18.3
urllib3==2.0.2
pywin32==306
customtkinter==5.1.3
python-vlc==3.0.18122
requests==2.30.0
pytube==15.0.0
pygame==2.5.0